Some sort of authorization added

+++ src/main/java/com/apea/training/parkWebsite/controller/requestHandler/plant/DeletePlantHandler.java
@@ -1,11 +1,13 @@
 package com.apea.training.parkWebsite.controller.requestHandler.plant;
 
 import com.apea.training.parkWebsite.controller.AppAssets;
+import com.apea.training.parkWebsite.controller.exception.AccessDeniedException;
 import com.apea.training.parkWebsite.controller.message.FrontMessageFactory;
 import com.apea.training.parkWebsite.controller.message.FrontendMessage;
 import com.apea.training.parkWebsite.controller.requestHandler.RequestHandler;
 import com.apea.training.parkWebsite.controller.utils.ControllerUtils;
 import com.apea.training.parkWebsite.domain.Plant;
+import com.apea.training.parkWebsite.domain.User;
 import com.apea.training.parkWebsite.service.PlantService;
 import com.apea.training.parkWebsite.service.impl.ServiceFactoryImpl;
 
@@ -19,6 +21,10 @@ public class DeletePlantHandler implements RequestHandler {
     @Override
     public String handle(HttpServletRequest request, HttpServletResponse response) {
         AppAssets assets = AppAssets.getInstance();
+
+        if (ControllerUtils.getCurrentUserId(request) == null) {return REDIRECT + assets.get("LOGIN_PAGE");}
+        if (ControllerUtils.getCurrentUserRole(request) == User.Role.FORESTER) {throw new AccessDeniedException("User is not the owner or a taskmaster");}
+
         List<FrontendMessage> generalMessages = new ArrayList<>();
         Integer areaId = deletePlant(request, generalMessages);
         ControllerUtils.saveGeneralMsgsInSession(request, generalMessages);
